pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "bicam"
version = "0.1.1"
description = "BICAM Dataset Downloader - Access comprehensive congressional data"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ryan Delano", email = "bicam.data@gmail.com"},
]
maintainers = [
    {name = "Ryan Delano", email = "bicam.data@gmail.com"},
]
keywords = ["congress", "legislation", "political-science", "dataset", "government-data"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "boto3>=1.26.0",
    "tqdm>=4.65.0",
    "pandas>=1.3.0",
    "polars>=0.20.0",
    "requests>=2.28.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=0.990",
    "moto>=5.0.28",
    "responses>=0.22.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
]

[project.scripts]
bicam = "bicam.cli:main"

[project.urls]
Homepage = "https://github.com/bicam-data/bicam"
Documentation = "https://bicam.readthedocs.io"
Repository = "https://github.com/bicam-data/bicam.git"
Issues = "https://github.com/bicam-data/bicam/issues"
Changelog = "https://github.com/bicam-data/bicam/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["bicam"]

[tool.setuptools.package-data]
bicam = ["py.typed"]

[tool.setuptools.exclude-package-data]
"*" = ["tests", "tests.*"]

[tool.setuptools.package-dir]
"" = "."

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "PT", "SIM", "PD"]
ignore = ["E501", "B008"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=bicam --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
source = ["bicam"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "black>=24.8.0",
    "moto>=5.0.28",
    "mypy>=1.14.1",
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.1",
    "responses>=0.25.7",
    "ruff>=0.12.2",
    "sphinx>=7.1.2",
    "sphinx-autodoc-typehints>=2.0.1",
    "sphinx-rtd-theme>=3.0.2",
]
